forked MailTypes: on CWrapMail and MailTypes
