+++ ChatServer/MailTypes.hpp
@@ -1,9 +1,6 @@
 #ifndef MAILTYPES_HPP
 #define MAILTYPES_HPP
 
-#include <string>
-#include <memory>
-
 enum MAIL_TYPE
 {
     MESSAGE
@@ -18,26 +15,6 @@ struct Mail
 {
     MAIL_TYPE typeMail;
     char data[1024];
-};
-
-class CWrapMail
-{
-    public:
-    CWrapMail(const Mail& mail,const int Id)
-        : typeMail(mail.typeMail)
-        , data(mail.data)
-        , clientId(Id)
-    {}
-
-    void getMail( Mail& mail) const
-    {
-        std::copy(data.begin(), data.end(), mail.data);
-        mail.typeMail = typeMail;
-    }
-
-    MAIL_TYPE typeMail;
-    std::string data;
-    const int clientId;
 };
 
 #endif // MAILTYPES_HPP
